5560 Apply multiple iterators in IteratorFactory
