+++ java/core/src/main/java/sleeper/core/iterator/IteratorFactory.java
@@ -15,7 +15,6 @@
  */
 package sleeper.core.iterator;
 
-import org.apache.commons.lang3.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -27,16 +26,8 @@ import sleeper.core.util.ObjectFactory;
 import sleeper.core.util.ObjectFactoryException;
 
 import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Locale;
 import java.util.Optional;
-import java.util.Set;
-import java.util.stream.Collectors;
-
-import static java.util.stream.Collectors.joining;
-import static java.util.stream.Collectors.toUnmodifiableSet;
 
 /**
  * A factory class for Sleeper iterators.
@@ -51,12 +42,9 @@ public class IteratorFactory {
     }
 
     /**
-     * Creates an initialises an iterator.
-     *
-     * The named iterator class is created and its {@link ConfigStringIterator#init(String, Schema)} method is called.
-     * If the special keyword for aggregation is specified as the class name
-     * {@link DataEngine#AGGREGATION_ITERATOR_NAME},
-     * then an aggregation iterator is created and initialised.
+     * Creates iterators based on the given configuration. This includes filtering, aggregation, and custom iterators.
+     * For custom iterators, the named iterator class is created and its
+     * {@link ConfigStringIterator#init(String, Schema)} method is called.
      *
      * @param  iteratorConfig            config object holding iterator details
      * @param  schema                    the Sleeper {@link Schema} of the {@link Row} objects
@@ -67,7 +55,19 @@ public class IteratorFactory {
      */
     public SortedRowIterator getIterator(IteratorConfig iteratorConfig, Schema schema) throws IteratorCreationException {
         try {
-            if (iteratorConfig.getFilteringString() == null && iteratorConfig.getAggregationString() == null) {
+            List<SortedRowIterator> iterators = new ArrayList<>();
+            if (iteratorConfig.getFilteringString() != null) {
+                for (AgeOffFilter filter : AgeOffFilter.parseConfig(iteratorConfig.getFilteringString())) {
+                    iterators.add(new AgeOffIterator(filter));
+                }
+            }
+            if (iteratorConfig.getAggregationString() != null) {
+                List<Aggregation> aggregations = Aggregation.parseConfig(iteratorConfig.getAggregationString(), schema);
+                List<String> groupingFields = schema.streamKeyFields().map(Field::getName).toList();
+                FilterAggregationConfig config = new FilterAggregationConfig(groupingFields, Optional.empty(), 0, aggregations);
+                iterators.add(new AggregationFilteringIterator(config, schema));
+            }
+            if (iteratorConfig.getIteratorClassName() != null) {
                 ConfigStringIterator iterator;
                 String className = iteratorConfig.getIteratorClassName();
 
@@ -80,131 +80,11 @@ public class IteratorFactory {
                 LOGGER.debug("Created iterator of class {}", className);
                 iterator.init(iteratorConfig.getIteratorConfigString(), schema);
                 LOGGER.debug("Initialised iterator with config {}", iteratorConfig.getIteratorConfigString());
-                return iterator;
-            } else {
-                return new AggregationFilteringIterator(getConfigFromProperties(iteratorConfig, schema), schema);
+                iterators.add(iterator);
             }
+            return new SortedRowIterators(iterators);
         } catch (RuntimeException | ObjectFactoryException exc) {
             throw new IteratorCreationException(exc);
-        }
-    }
-
-    /**
-     * Gets the filter aggregation config when set by table properties.
-     * Currently just gets the filter ageoff(column,age).
-     *
-     *
-     * @param  iteratorConfig config for an iterator that should have filters set in it
-     * @param  schema         the Sleeper {@link Schema} of the {@link Row} objects
-     * @return                filter aggregation config to be used in an iterator
-     */
-    private FilterAggregationConfig getConfigFromProperties(IteratorConfig iteratorConfig, Schema schema) throws IteratorCreationException {
-        List<String> groupingColumns = new ArrayList<>();
-        groupingColumns.addAll(schema.getRowKeyFieldNames());
-        groupingColumns.addAll(schema.getSortKeyFieldNames());
-        long maxAge = 0L;
-        String filterName = null;
-
-        if (iteratorConfig.getFilteringString() != null && !iteratorConfig.getFilteringString().equals("")) {
-            String[] filterParts = StringUtils.deleteWhitespace(iteratorConfig.getFilteringString()).split("\\(");
-            if ("ageoff".equals(filterParts[0].toLowerCase(Locale.ENGLISH))) {
-                String[] filterInput = StringUtils.chop(filterParts[1]).split(","); //Chop to remove the trailing ')'
-                filterName = filterInput[0];
-                maxAge = Long.parseLong(filterInput[1]);
-                groupingColumns.add(filterInput[0]);
-            } else {
-                throw new IllegalArgumentException("Sleeper table filter not set to match ageOff(column,age), was: " + filterParts[0]);
-            }
-        }
-        List<Aggregation> aggregations = List.of();
-        if (iteratorConfig.getAggregationString() != null && !iteratorConfig.getAggregationString().equals("")) {
-            aggregations = generateAggregationsFromProperty(iteratorConfig);
-            validateAggregations(aggregations, schema);
-        }
-
-        return new FilterAggregationConfig(groupingColumns,
-                filterName != null ? Optional.of(filterName) : Optional.empty(),
-                maxAge,
-                aggregations);
-    }
-
-    private List<Aggregation> generateAggregationsFromProperty(IteratorConfig iteratorConfig) {
-        List<Aggregation> aggregations = new ArrayList<Aggregation>();
-        String[] aggregationSplits = StringUtils.deleteWhitespace(iteratorConfig.getAggregationString()).split("\\),");
-        Arrays.stream(aggregationSplits).forEach(presentAggregation -> {
-            AggregationString aggObject = new AggregationString(presentAggregation
-                    .replaceAll("\\)", "")
-                    .split("\\("));
-            try {
-                aggregations.add(new Aggregation(aggObject.getColumnName(),
-                        AggregationOp.valueOf(aggObject.getOpName().toUpperCase(Locale.ENGLISH).replaceFirst("^MAP_", ""))));
-            } catch (RuntimeException e) {
-                throw new IllegalArgumentException("Unable to parse operand. Operand: " + aggObject.getOpName(), e);
-            }
-        });
-        return aggregations;
-    }
-
-    private void validateAggregations(List<Aggregation> aggregations, Schema schema) {
-        validateNoRowKeySortKeyAggregations(aggregations, schema);
-        validateNoDuplicateAggregations(aggregations);
-        validateAggregatedColumnsMatchValueColumns(aggregations, schema);
-    }
-
-    private void validateNoRowKeySortKeyAggregations(List<Aggregation> aggregations, Schema schema) {
-        Set<String> keyFields = schema.streamKeyFields()
-                .map(Field::getName)
-                .collect(toUnmodifiableSet());
-        List<Aggregation> rowKeySortKeyViolations = aggregations.stream().filter(agg -> keyFields.contains(agg.column())).toList();
-
-        if (!rowKeySortKeyViolations.isEmpty()) {
-            String errStr = rowKeySortKeyViolations.stream()
-                    .map(Aggregation::column)
-                    .collect(Collectors.joining(", "));
-            throw new IllegalArgumentException("Column for aggregation not allowed to be a Row Key or Sort Key. Column names: " + errStr);
-        }
-    }
-
-    private void validateNoDuplicateAggregations(List<Aggregation> aggregations) {
-        HashMap<String, Boolean> aggMap = new HashMap<String, Boolean>();
-        aggregations.stream().forEach(aggregation -> {
-            if (aggMap.putIfAbsent(aggregation.column(), Boolean.TRUE) != null) {
-                throw new IllegalArgumentException("Not allowed duplicate columns for aggregation. Column name: " + aggregation.column());
-            }
-        });
-    }
-
-    private void validateAggregatedColumnsMatchValueColumns(List<Aggregation> aggregations, Schema schema) {
-        Set<String> aggregationColumns = aggregations.stream().map(Aggregation::column).collect(toUnmodifiableSet());
-        Set<String> valueFields = schema.getValueFields().stream().map(Field::getName).collect(toUnmodifiableSet());
-
-        if (!aggregationColumns.containsAll(valueFields)) {
-            String errStr = schema.getValueFields().stream().map(Field::getName)
-                    .filter(presCol -> !aggregationColumns.contains(presCol))
-                    .collect(joining(", "));
-            throw new IllegalArgumentException("Not all value fields have aggregation declared. Missing columns: " + errStr);
-        }
-        if (!valueFields.containsAll(aggregationColumns)) {
-            String errStr = aggregations.stream().map(Aggregation::column)
-                    .filter(aggregationColumn -> !valueFields.contains(aggregationColumn))
-                    .collect(joining(", "));
-            throw new IllegalArgumentException("Not all aggregated fields are declared in the schema. Missing fields: " + errStr);
-        }
-    }
-
-    /**
-     * Record to provide parse object from the aggregation string for clarity.
-     *
-     * @param aggregationParts array contain sepeate parts of the aggregation.
-     */
-    private record AggregationString(String[] aggregationParts) {
-
-        String getOpName() {
-            return aggregationParts[0];
-        }
-
-        String getColumnName() {
-            return aggregationParts[1];
         }
     }
